# docs/swagger.json
{"consumes":["application/json","application/json"],"produces":["application/json","application/json","application/json","application/json","application/json"],"swagger":"2.0","info":{"description":"Delete an existing sample by its name","contact":{}},"paths":{"/samples":{"get":{"description":"Get all sample items","produces":["application/json"],"tags":["samples"],"summary":"List samples","responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/main.sample"}}}}},"post":{"description":"Add a new sample from the request body","consumes":["application/json"],"produces":["application/json"],"tags":["samples"],"summary":"Create a new sample","parameters":[{"description":"sample to add","name":"sample","in":"body","required":true,"schema":{"$ref":"#/definitions/main.sample"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/main.sample"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/gin.H"}},"409":{"description":"Conflict","schema":{"$ref":"#/definitions/gin.H"}}}}},"/samples/{name}":{"get":{"description":"Get a single sample by its name","produces":["application/json"],"tags":["samples"],"summary":"Get sample by Name","parameters":[{"type":"string","description":"sample Name","name":"name","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/main.sample"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/gin.H"}}}},"put":{"description":"Update an existing sample by its name","consumes":["application/json"],"produces":["application/json"],"tags":["samples"],"summary":"Update sample by Name","parameters":[{"type":"string","description":"sample Name","name":"name","in":"path","required":true},{"description":"Updated sample Data","name":"sample","in":"body","required":true,"schema":{"$ref":"#/definitions/main.sample"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/main.sample"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/gin.H"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/gin.H"}}}},"delete":{"description":"Delete an existing sample by its name","produces":["application/json"],"tags":["samples"],"summary":"Delete sample by Name","parameters":[{"type":"string","description":"sample Name","name":"name","in":"path","required":true}],"responses":{"200":{"description":"sample deleted message","schema":{"$ref":"#/definitions/gin.H"}},"404":{"description":"sample not found message","schema":{"$ref":"#/definitions/gin.H"}}}}}},"definitions":{"gin.H":{"type":"object","additionalProperties":{}},"main.sample":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"title":{"type":"string"}}}}}